refactor: add abstract class and add dealer
