+++ src/main/kotlin/blackjack/domain/player/Dealer.kt
@@ -1,0 +1,14 @@
+package blackjack.domain.player
+
+import blackjack.domain.Card
+
+class Dealer : AbstractPlayer() {
+
+    override fun drawCard(newCard: Card) {
+        TODO("Not yet implemented")
+    }
+
+    override fun isBust(): Boolean {
+        return calculateCard() > 16
+    }
+}
